+++ tests/test_confirmable.py
@@ -201,9 +201,7 @@ def test_expired_confirmation_token(client, get_message):
 
 
 @pytest.mark.registerable()
-def test_email_conflict_for_confirmation_token(
-    app, client, get_message, sqlalchemy_datastore
-):
+def test_email_conflict_for_confirmation_token(app, client, get_message):
     with capture_registrations() as registrations:
         data = dict(email="mary@lp.com", password="awesome sunset", next="")
         client.post("/register", data=data, follow_redirects=True)
@@ -214,8 +212,8 @@ def test_email_conflict_for_confirmation_token(
     # Change the user's email
     user.email = "tom@lp.com"
     with app.app_context():
-        sqlalchemy_datastore.put(user)
-        sqlalchemy_datastore.commit()
+        app.security.datastore.put(user)
+        app.security.datastore.commit()
 
     response = client.get("/confirm/" + token, follow_redirects=True)
     msg = get_message("INVALID_CONFIRMATION_TOKEN")